backend/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
Flask==2.0.1
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.1
protobuf==3.20.3
pycparser==2.21
pytz==2022.7.1
six==1.16.0
Werkzeug==2.2.2
zope.interface==5.5.2
